Using your Nodal!
Implemented methods of adding nodes to our graph.c. When a new tile is
placed on the board, nodes are added to each of the end points of a tile
and are used to connect various tiles together. Each node connect two
tiles at either a vertial or horizontal connection.